Move System class to a new submodule openep.view.system_manager
